refact: Move methods to `FieldMethods` trait
